chore: move types for no circular dependences
