+++ src/common/Form/ImageInput/setSrc.ts
@@ -1,6 +1,6 @@
 import type { IConvertedFileMeta } from 'src/types'
 import type { IUploadedFileMeta } from '../../../stores/storage'
-import type { IInputValue } from './ImageInput'
+import type { IInputValue } from './types'
 
 export const setSrc = (file: IInputValue): string => {
   if (!file) return ''
