+++ src/common/Form/ImageInput/types.ts
@@ -1,0 +1,6 @@
+import type { IConvertedFileMeta } from 'src/types'
+import type { IUploadedFileMeta } from '../../../stores/storage'
+
+export type IInputValue = IUploadedFileMeta | IConvertedFileMeta
+export type IMultipleInputValue = IInputValue[]
+export type IValue = IInputValue | IMultipleInputValue
